refactor: ServiceImpl 클래스 제거 및 service 구조 단순화
