+++ src/main/java/roomescape/controller/ReservationController.java
@@ -13,7 +13,6 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.application.dto.ReservationRequest;
 import roomescape.application.dto.ReservationResponse;
 import roomescape.application.service.ReservationService;
-import roomescape.application.service.ReservationServiceImpl;
 
 @RestController
 @RequestMapping("reservations")
@@ -21,7 +20,7 @@ public class ReservationController {
 
     private final ReservationService reservationService;
 
-    public ReservationController(ReservationServiceImpl reservationService) {
+    public ReservationController(ReservationService reservationService) {
         this.reservationService = reservationService;
     }
